+++ biodumpy/__init__.py
@@ -1,4 +1,4 @@
 from biodumpy.biodumpy import Biodumpy
 from biodumpy.biodumpy import BiodumpyException
 from .input import Input
-# from .input import *
+# from .utils import *