+++ biodumpy/inputs/BOLD.py
@@ -12,18 +12,13 @@ class BOLD(Input):
 	query : list
 	    The list of taxa to query.
 	summary : bool, optional
-	    If True, the function returns a summary of the downloaded metadata instead of the full records.
-		Default is False.
-	fasta : bool, optional
-		If True, the function downloads the FASTA file.
-		Default is False.
+	    If True, the function returns a summary of the downloaded metadata instead of the full records. Default is False.
 	bulk : bool, optional
 		If True, the function creates a bulk file.
-		For further information, see the documentation of the Biodumpy package.
-		Default is False.
+		For further information, see the documentation of the Biodumpy package. Default is False.
 	output_format : string, optional
 		The format of the output file.
-		The options available are: 'json', 'fasta', 'pdf'.
+		The options available are: 'json', 'fasta'.
 		Default is 'json'.
 
 	Details
@@ -52,19 +47,20 @@ class BOLD(Input):
 	>>> bdp.start(taxa, output_path='./downloads/{date}/{module}/{name}')
 	"""
 
-	def __init__(self, summary: bool = False, fasta: bool = False, output_format: str = "json", bulk: bool = False):
+	def __init__(self, summary: bool = False, output_format: str = "json", bulk: bool = False):
 		super().__init__(output_format, bulk)
 		self.summary = summary
-		self.fasta = fasta
+		# self.fasta = fasta
 
-		if self.fasta and output_format != "fasta":
-			raise ValueError("Invalid output_format. Expected fasta.")
+		# if self.fasta and output_format != "fasta":
+		# 	raise ValueError("Invalid output_format. Expected fasta.")
 
 		if output_format not in {"json", "fasta"}:
 			raise ValueError('Invalid output_format. Expected "json" or "fasta".')
 
 	def _download(self, query, **kwargs) -> list:
-		if self.fasta:
+		# if self.fasta:
+		if self.output_format == "fasta":
 			response = requests.get(f"http://v4.boldsystems.org/index.php/API_Public/sequence?taxon={query}")
 
 			if response.status_code != 200: