+++ biodumpy/inputs/GBIF.py
@@ -5,7 +5,7 @@ from biodumpy import Input, BiodumpyException
 
 class GBIF(Input):
 	"""
-	Query the GBIF database to retrieve taxon data.
+	Query the GBIF database to retrieve taxonomic and occurrence data.
 
 	Parameters
 	----------
@@ -14,17 +14,29 @@ class GBIF(Input):
 	dataset_key : str
 	    GBIF dataset key. The default is set to the GBIF Backbone Taxonomy dataset key.
 	limit : int
-	    The maximum number of names to retrieve from the taxonomy backbone for a taxon. Default is 20.
+	    The maximum number of names to retrieve from the taxonomy backbone for a taxon.
+	    Default is 20.
 	accepted_only : bool, optional
-	    If True, the function returns only the accepted name. Default is True.
+	    If True, the function returns only the accepted name. See the Details section for more information.
+	    Default is True.
 	occ : bool, optional
-	    If True, the function also returns the occurrences of a taxon. Default is False.
+	    If True, the function also returns the occurrences of a taxon. See the Details section for more information.
+	    Default is False.
 	geometry : str, optional
-	    A spatial polygon to filter occurrences within a specified area. Default is an empty string.
+	    A spatial polygon to filter occurrences within a specified area.
+	    Default is an empty string.
 	bulk : bool, optional
-	    If True, the function creates a bulk file. For further information, see the documentation of the Biodumpy package. Default is False.
+	    If True, the function creates a bulk file. For further information, see the documentation of the Biodumpy package.
+	    Default is False.
 	output_format : str, optional
-	    The format of the output file. The options available are: 'json', 'fasta', 'pdf'. Default is 'json'.
+	    The format of the output file. The options available are: 'json', 'fasta', 'pdf'.
+	    Default is 'json'.
+
+	Details
+	-------
+	If accepted_only and occ are both set to True, the function returns the occurrences of the taxon, including all synonyms. If accepted_only is set to False and occ is set to True, the function also downloads occurrences for lower taxonomic levels.
+	Refer to the acceptedTaxonKey and taxonKey parameters of the GBIF API endpoint at https://api.gbif.org/v1/occurrence/search for further details.
+
 
 	Example
 	-------
@@ -33,7 +45,7 @@ class GBIF(Input):
 	# GBIF dataset key
 	>>> gbif_backbone = 'd7dddbf4-2cf0-4f39-9b2a-bb099caae36c'
 	# Taxa list
-	>>> taxa = ['Alytes muletensis (Sanchíz & Adrover, 1979)', 'Bufotes viridis (Laurenti, 1768)']
+	>>> taxa = ['Alytes muletensis', 'Bufotes viridis']
 	# Set the module and start the download
 	>>> bdp = Biodumpy([GBIF(dataset_key=gbif_backbone, limit=20, accepted_only=True, occ=False, bulk=False, output_format='json')])
 	>>> bdp.start(taxa, output_path='./downloads/{date}/{module}/{name}')
@@ -59,15 +71,10 @@ class GBIF(Input):
 		if output_format != "json":
 			raise ValueError('Invalid output_format. Expected "json".')
 
-		# if occ and not accepted_only:
-		# 	raise ValueError("Invalid accepted_only. Expected True.")
-
 	def _download(self, query, **kwargs) -> list:
 		payload = []
 
 		# Search taxonomy
-		# response = requests.get(f"https://api.gbif.org/v1/species/search?datasetKey={self.dataset_key}&q={query}&limit={self.limit}")
-
 		response = requests.get(f"https://api.gbif.org/v1/species?",
 		                        params={"datasetKey": self.dataset_key,
 		                                "name": query,
@@ -78,26 +85,55 @@ class GBIF(Input):
 		if response.status_code != 200:
 			raise BiodumpyException(f"Taxonomy request. Error {response.status_code}")
 
+		# if len(payload) == 0:
+		# 	raise BiodumpyException(f"Taxonomy not found.")
+
 		if response.content:
 
 			payload = response.json()["results"]
 
-			if self.accepted:
-				# We keep the record only if the query corresponds to the scientific name in the data downloaded.
-				# payload = list(filter(lambda x: x.get("taxonomicStatus") == "ACCEPTED" and str(query[0]) in x.get("scientificName", ""), payload))
+			if len(payload) > 1:
+				# raise BiodumpyException(f"Multiple equal matches for {query}")
+				keys = [entry["key"] for entry in payload]
 
+				# Display the list of taxon ID options
+				options = [{"index": i, "taxon_id": keys} for i, keys in enumerate(keys)]
+				print(f"\nOptions for {query}:\n")
+				for option in options:
+					print(f"Taxon ID: {option['taxon_id']} - Index: {option['index']}.")
+				print(f"{len(keys) + 1}. Skip")
+
+				# Get user selection
+				choice = input("Please enter the index corresponding to the correct taxon GBIF ID or choose 'Skip': ")
+
+				# Process selection
+				try:
+					choice = int(choice)
+					if 0 <= choice < len(keys):
+						payload = [payload[choice]]
+					else:
+						print("Skipped selection.")
+				except ValueError:
+					print("Invalid input, please enter a number.")
+
+
+			if self.accepted:
 				if payload[0].get("taxonomicStatus") != "ACCEPTED":
 					acceptedKey = payload[0].get("acceptedKey")
 					response_accepted = requests.get(f"https://api.gbif.org/v1/species/{acceptedKey}")
 					payload = response_accepted.json()
+				else:
+					acceptedKey = payload[0].get("key")
 
 				if self.occ and len(payload) > 0:
+				# A taxon key from the GBIF backbone. Only synonym taxa are included in the search, so a search for Aves with acceptedTaxonKey=212 (i.e. /occurrence/search?taxonKey=212) will match occurrences identified as birds, but not any known family, genus or species of bird.Parameter may be repeated.
 					payload = self._download_gbif_occ(accepted_taxon_key=acceptedKey, geometry=self.geometry)
 
 			else:
 				payload = payload[0]
 
 				if self.occ and len(payload) > 0:
+					# A taxon key from the GBIF backbone. All included (child) and synonym taxa are included in the search, so a search for Aves with taxonKey=212 (i.e. /occurrence/search?taxonKey=212) will match all birds, no matter which species.Parameter may be repeated.
 					payload = self._download_gbif_occ(taxon_key=payload["key"], geometry=self.geometry)
 
 		return payload
@@ -114,6 +150,9 @@ class GBIF(Input):
 
 		if response_occ.status_code != 200:
 			raise BiodumpyException(f"Occurrence request. Error {response_occ.status_code}")
+
+		if response_occ.status_code == 0:
+			raise BiodumpyException(f"Occurrence not found.")
 
 		if response_occ.content:
 			payload_occ = response_occ.json()