+++ tests/BOLD.py
@@ -13,14 +13,14 @@ from biodumpy.inputs import BOLD
 trap = io.StringIO()
 
 
-def bold_query(query, summary, fasta, output_format):
+def bold_query(query, summary, output_format):
 	# Create temporary directory
 	with tempfile.TemporaryDirectory() as temp_dir:
 		# Construct the dynamic path using formatted strings
 		dynamic_path = os.path.join(temp_dir)
 
 	# Start biodumpy function
-	bdp = Biodumpy([BOLD(bulk=True, summary=summary, fasta=fasta, output_format=output_format)])
+	bdp = Biodumpy([BOLD(bulk=True, summary=summary, output_format=output_format)])
 	bdp.start(elements=query, output_path=f"{dynamic_path}/downloads/{{date}}/{{module}}/{{name}}")
 
 	# Retrieve a file path
@@ -45,14 +45,8 @@ def test_bold_initialization():
 
 	# Verify default parameters
 	assert bold.summary == False
-	assert bold.fasta == False
 	assert bold.bulk == False
 	assert bold.output_format == "json"
-
-	# Objective: Verify that the class raises a ValueError when an invalid value is provided for the
-	# output_format parameter. If fasta parameter is True, output_format has to be 'fasta'.
-	with pytest.raises(ValueError, match="Invalid output_format. Expected fasta."):
-		BOLD(output_format="json", fasta=True)
 
 	# Objective: Verify that the class raises a ValueError when an invalid value is provided for the
 	# output_format parameter.
@@ -61,21 +55,21 @@ def test_bold_initialization():
 
 
 @pytest.mark.parametrize(
-	"query, summary, fasta, output_format",
+	"query, summary, output_format",
 	[
-		(["Alytes muletensis"], True, False, "json"),
-		(["Alytes muletensis"], False, False, "json"),
-		(["Alytes muletensis"], False, True, "fasta")
+		(["Alytes muletensis"], True, "json"),
+		(["Alytes muletensis"], False, "json"),
+		(["Alytes muletensis"], False, "fasta")
 	],
 )
-def test_download(query, summary, fasta, output_format):
+def test_download(query, summary, output_format):
 	with redirect_stdout(trap):
-		data = bold_query(query=query, summary=summary, fasta=fasta, output_format=output_format)
+		data = bold_query(query=query, summary=summary, output_format=output_format)
 
 	# Check if data is not empty
 	assert len(data) > 0, "data length is 0"
 
-	if summary is False and fasta is False:
+	if summary is False and output_format != "fasta":
 		# Check the main info in a BOLD JSON file
 		data = data[0]
 
@@ -133,7 +127,7 @@ def test_download(query, summary, fasta, output_format):
 		assert 'nucleotides' in seq, "nucleotides is not in sequences"
 
 
-	if summary and fasta is False:
+	if summary and output_format != "fasta":
 		# Check the summary structure
 		assert "record_id" in data[0], "record_id is not in data"
 		assert "processid" in data[0], "processid is not in data"
@@ -147,6 +141,6 @@ def test_download(query, summary, fasta, output_format):
 		assert "markercode" in data[0], "markercode is not in data"
 		assert "genbank_accession" in data[0], "genbank_accession is not in data"
 
-	if summary is False and fasta is True:
+	if summary is False and output_format == "fasta":
 		# Check if the fasta file starts with >
 		assert data.startswith(">")