+++ tests/GBIF.py
@@ -30,8 +30,8 @@ def gbif_query(query, accepted_only, occ, geometry):
 				accepted_only=accepted_only,
 				occ=occ,
 				geometry=geometry,
-				bulk=True,
-				output_format="json"
+				output_format="json",
+				bulk = True
 			)
 		]
 	)
@@ -64,11 +64,6 @@ def test_gbif_initialization():
 	assert gbif.output_format == "json"
 
 	# Objective: Verify that the class raises a ValueError when an invalid value is provided for the
-	# accepted_only parameter.
-	with pytest.raises(ValueError, match="Invalid accepted_only. Expected True."):
-		GBIF(occ=True, accepted_only=False)
-
-	# Objective: Verify that the class raises a ValueError when an invalid value is provided for the
 	# output_format parameter.
 	with pytest.raises(ValueError, match='Invalid output_format. Expected "json".'):
 		GBIF(output_format="xml")
@@ -77,8 +72,8 @@ def test_gbif_initialization():
 @pytest.mark.parametrize(
 	"query, accepted_only, occ, geometry",
 	[
-		(["Alytes muletensis (Sanchíz & Adrover, 1979)"], True, False, None),
-		(["Alytes muletensis (Sanchíz & Adrover, 1979)"], True, True,
+		(["Alytes muletensis"], True, False, None),
+		(["Alytes muletensis"], True, True,
 		 "POLYGON((0.248 37.604, 6.300 37.604, 6.300 41.472, 0.248 41.472, 0.248 37.604))"),
 	],
 )
@@ -97,11 +92,9 @@ def test_download(query, accepted_only, occ, geometry):
 		assert "nameKey" in data, "nameKey is not in data"
 		assert data["nameKey"] == 497717, "nameKey is not 497717"
 		assert "datasetKey" in data, "datasetKey is not in data"
-		assert data[
-			       "datasetKey"] == 'd7dddbf4-2cf0-4f39-9b2a-bb099caae36c', "datasetKey is not d7dddbf4-2cf0-4f39-9b2a-bb099caae36c"
+		assert data["datasetKey"] == 'd7dddbf4-2cf0-4f39-9b2a-bb099caae36c', "datasetKey is not d7dddbf4-2cf0-4f39-9b2a-bb099caae36c"
 		assert "constituentKey" in data, "constituentKey is not in data"
-		assert data[
-			       "constituentKey"] == '7ddf754f-d193-4cc9-b351-99906754a03b', "datasetKey is not 7ddf754f-d193-4cc9-b351-99906754a03b"
+		assert data["constituentKey"] == '7ddf754f-d193-4cc9-b351-99906754a03b', "datasetKey is not 7ddf754f-d193-4cc9-b351-99906754a03b"
 		assert "nubKey" in data, "nubKey is not in data"
 		assert data["nubKey"] == 2426609, "nubKey is not 2426609"
 		assert "parentKey" in data, "parentKey is not in data"
@@ -111,8 +104,7 @@ def test_download(query, accepted_only, occ, geometry):
 		assert "basionymKey" in data, "basionymKey is not in data"
 		assert data["basionymKey"] == 4409471, "basionymKey is not 4409471"
 		assert "basionym" in data, "basionym is not in data"
-		assert data[
-			       "basionym"] == 'Baleaphryne muletensis Sanchíz & Adrover, 1979', "basionym is not Baleaphryne muletensis Sanchíz & Adrover, 1979"
+		assert data["basionym"] == 'Baleaphryne muletensis Sanchíz & Adrover, 1979', "basionym is not Baleaphryne muletensis Sanchíz & Adrover, 1979"
 		assert "kingdom" in data, "kingdom is not in data"
 		assert data["kingdom"] == 'Animalia', "kingdom is not Animalia"
 		assert "phylum" in data, "phylum is not in data"
@@ -144,8 +136,7 @@ def test_download(query, accepted_only, occ, geometry):
 		assert "speciesKey" in data, "speciesKey is not in data"
 		assert data["speciesKey"] == 2426609, "speciesKey is not 2426609"
 		assert "scientificName" in data, "scientificName is not in data"
-		assert data[
-			       "scientificName"] == 'Alytes muletensis (Sanchíz & Adrover, 1979)', "scientificName is not Alytes muletensis (Sanchíz & Adrover, 1979)"
+		assert data["scientificName"] == 'Alytes muletensis (Sanchíz & Adrover, 1979)', "scientificName is not Alytes muletensis (Sanchíz & Adrover, 1979)"
 		assert "canonicalName" in data, "canonicalName is not in data"
 		assert data["canonicalName"] == "Alytes muletensis", "canonicalName is not Alytes muletensis"
 		assert "authorship" in data, "authorship is not in data"
@@ -160,16 +151,16 @@ def test_download(query, accepted_only, occ, geometry):
 		assert data["origin"] == 'SOURCE', "origin is not SOURCE"
 		assert "numDescendants" in data, "numDescendants is not in data"
 		assert data["numDescendants"] == 1, "numDescendants is not 1"
-		assert "numOccurrences" in data, "numOccurrences is not in data"
-		assert data["numOccurrences"] == 0, "numOccurrences is not 0"
-		assert "extinct" in data, "extinct is not in data"
-		assert "habitats" in data, "habitats is not in data"
+		# assert "numOccurrences" in data, "numOccurrences is not in data"
+		# assert data["numOccurrences"] == 0, "numOccurrences is not 0"
+		# assert "extinct" in data, "extinct is not in data"
+		# assert "habitats" in data, "habitats is not in data"
 		assert 'nomenclaturalStatus' in data, "nomenclaturalStatus is not in data"
-		assert 'threatStatuses' in data, "threatStatuses is not in data"
-		assert 'descriptions' in data, "descriptions is not in data"
-		assert 'vernacularNames' in data, "vernacularNames is not in data"
-		assert 'higherClassificationMap' in data, "higherClassificationMap is not in data"
-		assert 'synonym' in data, "synonym is not in data"
+		# assert 'threatStatuses' in data, "threatStatuses is not in data"
+		# assert 'descriptions' in data, "descriptions is not in data"
+		# assert 'vernacularNames' in data, "vernacularNames is not in data"
+		# assert 'higherClassificationMap' in data, "higherClassificationMap is not in data"
+		# assert 'synonym' in data, "synonym is not in data"
 	else:
 		assert "key" in data, "key is not in data"
 		assert "datasetKey" in data, "datasetKey is not in data"